+++ packages/tsurlfilter/CHANGELOG.md
@@ -16,10 +16,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Scriptlet exception rules with comma-containing arguments now work correctly
+  ([AdguardBrowserExtension#3533]).
 - Expanded wildcard TLD domains in DNR conversion for `$domain` and `$to`
-  modifiers [AdguardTeam/tsurlfilter#189].
+  modifiers [#189].
 
-[AdguardTeam/tsurlfilter#189]: https://github.com/AdguardTeam/tsurlfilter/issues/189
+[AdguardBrowserExtension#3533]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3533
+[#189]: https://github.com/AdguardTeam/tsurlfilter/issues/189
 
 ## [v5.0.1] - 2026-04-24